fix the float error due to division

+++ csi/volumeutils.py
@@ -773,7 +773,7 @@ def generate_client_volfile(volname):
         count = 3
         if data["type"] == "Replica2":
             count = 2
-        for i in range(0, (len(data["bricks"]) / count)):
+        for i in range(0, int(len(data["bricks"]) / count)):
             data["dht_subvol"].append("%s-replica-%d" % (data["volname"], i))
 
     template_file_path = os.path.join(
